chore: add course doubt section
